+++ src/check.ts
@@ -1,5 +1,6 @@
 import { isIn } from "@dwtechs/checkard";
 import type { MatchMode, MappedType } from "./types";
+import { COMPARATORS } from "./filter/map/comparator";
 
 const matchModes = {
   string: ["startsWith", "contains", "endsWith", "notContains", "equals", "notEquals", "lt", "lte", "gt", "gte", "in", "notIn"],
@@ -8,7 +9,7 @@ const matchModes = {
 };
 
 function matchMode(type: MappedType, matchMode: MatchMode): boolean {
-  return isIn(matchModes[type], matchMode);
+  return isIn([...COMPARATORS], matchMode) || isIn(matchModes[type], matchMode);
 }
 
 export {